Move provable model to model file + add doc comments
