+++ zkml/src/iop/verifier.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use crate::{
-    commit::{self, precommit}, iop::{context::ShapeStep, ChallengeStorage}, layers::{provable::{NodeCtx, NodeId, ProvableOpError, ToIterator, VerifiableCtx}, LayerProof}, lookup::{context::TableType, logup_gkr::verifier::verify_logup_proof}, tensor::Tensor, try_unzip, Claim, VectorTranscript
+    commit::{self, precommit}, iop::{context::ShapeStep, ChallengeStorage}, layers::{provable::{NodeCtx, NodeId, ProvableOpError, VerifiableCtx}, LayerProof}, lookup::{context::TableType, logup_gkr::verifier::verify_logup_proof}, model::ToIterator, tensor::Tensor, try_unzip, Claim, VectorTranscript
 };
 use anyhow::{anyhow, ensure};
 use ff_ext::ExtensionField;
