+++ zkml/src/padding.rs
@@ -3,15 +3,13 @@ use std::collections::HashMap;
 use anyhow::{Context, Result, anyhow, ensure};
 
 use crate::{
-    Element,
     layers::{
         convolution::Convolution,
         dense::Dense,
         flatten::Flatten,
         pooling::Pooling,
-        provable::{NodeId, ProvableModel, ProvableNode, ToIterator},
-    },
-    onnx_parse::{check_filter, safe_conv2d_shape, safe_maxpool2d_shape},
+        provable::{NodeId, ProvableNode},
+    }, model::{ProvableModel, ToIterator}, onnx_parse::{check_filter, safe_conv2d_shape, safe_maxpool2d_shape}, Element
 };
 type GarbagePad = Option<(Vec<usize>, Vec<usize>)>;
 type Shape = Vec<usize>;
